feat(api): add `initPayment`

+++ library/src/commonMain/kotlin/ink/literate/turbawself/decoders/payment.kt
@@ -13,7 +13,7 @@ fun decodePayment(payment: JsonObject): Payment {
         date = LocalDateTime.parse(payment["date"]!!.jsonPrimitive.content.dropLast(1)),
         updateDate = if (payment["dateMaj"] != null) LocalDateTime.parse(payment["dateMaj"]!!.jsonPrimitive.content.dropLast(1)) else null,
         type = payment["type"]!!,
-        amount = payment["montant"]!!.jsonPrimitive.double,
+        amount = payment["montant"]!!.jsonPrimitive.int,
         status = decodePaymentStatus(payment["statut"]!!.jsonPrimitive.content),
         transactionID = payment["idTransaction"]!!.jsonPrimitive.content,
         token = paymentToken,
